EXT-1622: moving to pages
