+++ app-nuxtjs-3-starter/pages/401.vue
@@ -1,0 +1,1 @@
+<template><span>401 page</span></template>
